@@ -1,96 +1,135 @@
 using Microsoft.AspNetCore.Mvc;
+
 using TrilhaApiDesafio.Context;
 using TrilhaApiDesafio.Models;
 
-namespace TrilhaApiDesafio.Controllers
+namespace TrilhaApiDesafio.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class TarefaController : ControllerBase
 {
-    [ApiController]
-    [Route("[controller]")]
-    public class TarefaController : ControllerBase
+    private readonly OrganizadorContext _context;
+
+    public TarefaController(OrganizadorContext context)
     {
-        private readonly OrganizadorContext _context;
+        _context = context;
+    }
 
-        public TarefaController(OrganizadorContext context)
-        {
-            _context = context;
-        }
+    [HttpGet("{id}")]
+    public IActionResult ObterPorId(int id)
+    {
+        // TODO: Buscar o Id no banco utilizando o EF
+        // TODO: Validar o tipo de retorno. Se não encontrar a tarefa, retornar NotFound,
+        // caso contrário retornar OK com a tarefa encontrada
+        // CONCLUÍDO
+        var tarefa = _context.Tarefas.Find(id);
+        if(tarefa == null)
+            return NotFound();
 
-        [HttpGet("{id}")]
-        public IActionResult ObterPorId(int id)
-        {
-            // TODO: Buscar o Id no banco utilizando o EF
-            // TODO: Validar o tipo de retorno. Se não encontrar a tarefa, retornar NotFound,
-            // caso contrário retornar OK com a tarefa encontrada
-            return Ok();
-        }
+        return Ok(tarefa);
+    }
 
-        [HttpGet("ObterTodos")]
-        public IActionResult ObterTodos()
-        {
-            // TODO: Buscar todas as tarefas no banco utilizando o EF
-            return Ok();
-        }
 
-        [HttpGet("ObterPorTitulo")]
-        public IActionResult ObterPorTitulo(string titulo)
-        {
-            // TODO: Buscar  as tarefas no banco utilizando o EF, que contenha o titulo recebido por parâmetro
-            // Dica: Usar como exemplo o endpoint ObterPorData
-            return Ok();
-        }
+    [HttpGet("ObterTodos")]
+    public IActionResult ObterTodos()
+    {
+        // TODO: Buscar todas as tarefas no banco utilizando o EF
+        // CONCLUÍDO
+        var tarefa = _context.Tarefas.ToList();
+        if(tarefa == null)
+            return NotFound();
 
-        [HttpGet("ObterPorData")]
-        public IActionResult ObterPorData(DateTime data)
-        {
-            var tarefa = _context.Tarefas.Where(x => x.Data.Date == data.Date);
-            return Ok(tarefa);
-        }
+        return Ok(tarefa);
+    }
 
-        [HttpGet("ObterPorStatus")]
-        public IActionResult ObterPorStatus(EnumStatusTarefa status)
-        {
-            // TODO: Buscar  as tarefas no banco utilizando o EF, que contenha o status recebido por parâmetro
-            // Dica: Usar como exemplo o endpoint ObterPorData
-            var tarefa = _context.Tarefas.Where(x => x.Status == status);
-            return Ok(tarefa);
-        }
 
-        [HttpPost]
-        public IActionResult Criar(Tarefa tarefa)
-        {
-            if (tarefa.Data == DateTime.MinValue)
-                return BadRequest(new { Erro = "A data da tarefa não pode ser vazia" });
+    [HttpGet("ObterPorTitulo")]
+    public IActionResult ObterPorTitulo(string titulo)
+    {
+        // TODO: Buscar  as tarefas no banco utilizando o EF, que contenha o titulo recebido por parâmetro
+        // Dica: Usar como exemplo o endpoint ObterPorData
+        // CONCLUÍDO
+        var tarefa = _context.Tarefas.Where(x => x.Titulo.Contains(titulo));
+        if(tarefa == null)
+            return NotFound();
 
-            // TODO: Adicionar a tarefa recebida no EF e salvar as mudanças (save changes)
-            return CreatedAtAction(nameof(ObterPorId), new { id = tarefa.Id }, tarefa);
-        }
+        return Ok(tarefa);
+    }
 
-        [HttpPut("{id}")]
-        public IActionResult Atualizar(int id, Tarefa tarefa)
-        {
-            var tarefaBanco = _context.Tarefas.Find(id);
 
-            if (tarefaBanco == null)
-                return NotFound();
+    [HttpGet("ObterPorData")]
+    public IActionResult ObterPorData(DateTime data)
+    {
+        var tarefa = _context.Tarefas.Where(x => x.Data.Date == data.Date);
+        return Ok(tarefa);
+    }
 
-            if (tarefa.Data == DateTime.MinValue)
-                return BadRequest(new { Erro = "A data da tarefa não pode ser vazia" });
 
-            // TODO: Atualizar as informações da variável tarefaBanco com a tarefa recebida via parâmetro
-            // TODO: Atualizar a variável tarefaBanco no EF e salvar as mudanças (save changes)
-            return Ok();
-        }
+    [HttpGet("ObterPorStatus")]
+    public IActionResult ObterPorStatus(EnumStatusTarefa status)
+    {
+        // TODO: Buscar  as tarefas no banco utilizando o EF, que contenha o status recebido por parâmetro
+        // Dica: Usar como exemplo o endpoint ObterPorData
+        var tarefa = _context.Tarefas.Where(x => x.Status == status);
+        return Ok(tarefa);
+    }
 
-        [HttpDelete("{id}")]
-        public IActionResult Deletar(int id)
-        {
-            var tarefaBanco = _context.Tarefas.Find(id);
 
-            if (tarefaBanco == null)
-                return NotFound();
+    [HttpPost]
+    public IActionResult Criar(Tarefa tarefa)
+    {
+        if(tarefa.Data == DateTime.MinValue)
+            return BadRequest(new { Erro = "A data da tarefa não pode ser vazia" });
 
-            // TODO: Remover a tarefa encontrada através do EF e salvar as mudanças (save changes)
-            return NoContent();
-        }
+        // TODO: Adicionar a tarefa recebida no EF e salvar as mudanças (save changes)
+        // CONCLUÍDO
+        _context.Add(tarefa);
+        _context.SaveChanges();
+
+        return CreatedAtAction(nameof(ObterPorId), new { id = tarefa.Id }, tarefa);
+    }
+
+
+    [HttpPut("{id}")]
+    public IActionResult Atualizar(int id, Tarefa tarefa)
+    {
+        var tarefaBanco = _context.Tarefas.Find(id);
+
+        if(tarefaBanco == null)
+            return NotFound();
+
+        if(tarefa.Data == DateTime.MinValue)
+            return BadRequest(new { Erro = "A data da tarefa não pode ser vazia" });
+
+        // TODO: Atualizar as informações da variável tarefaBanco com a tarefa recebida via parâmetro
+        // TODO: Atualizar a variável tarefaBanco no EF e salvar as mudanças (save changes)
+        // CONCLUÍDO
+
+        tarefaBanco.Titulo = tarefa.Titulo;
+        tarefaBanco.Descricao = tarefa.Descricao;
+        tarefaBanco.Data = tarefa.Data;
+        tarefaBanco.Status = tarefa.Status;
+
+        _context.Tarefas.Update(tarefaBanco);
+        _context.SaveChanges();
+        return Ok(tarefaBanco);
+    }
+
+
+    [HttpDelete("{id}")]
+    public IActionResult Deletar(int id)
+    {
+        var tarefaBanco = _context.Tarefas.Find(id);
+
+        if(tarefaBanco == null)
+            return NotFound();
+
+        // TODO: Remover a tarefa encontrada através do EF e salvar as mudanças (save changes)
+        // CONCLUÍDO
+        _context.Tarefas.Remove(tarefaBanco);
+        _context.SaveChanges();
+
+        return NoContent();
     }
 }
