@@ -1,11 +1,14 @@
 using System.Text.Json.Serialization;
+
 using Microsoft.EntityFrameworkCore;
+
 using TrilhaApiDesafio.Context;
+
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-builder.Services.AddDbContext<OrganizadorContext>(options =>
+builder.Services.AddDbContext<OrganizadorContext>(options => 
     options.UseSqlServer(builder.Configuration.GetConnectionString("ConexaoPadrao")));
 
 builder.Services.AddControllers().AddJsonOptions(options =>
